Fix more layout bugs, bugs with search and remove introduction warning
