+++ apps/docs/app/features/portable-text/PortableText.tsx
@@ -264,6 +264,20 @@ const components: Partial<PortableTextReactComponents> = {
       }
       return <CodeBlock code={value.reactImport} marginTop={3} />;
     },
+    introduction: ({ value }) => (
+      <Box marginBottom={8}>
+        <PortableText
+          value={value.introduction}
+          components={{
+            block: {
+              normal: ({ children }: any) => (
+                <Text variant="md">{children}</Text>
+              ),
+            },
+          }}
+        />
+      </Box>
+    ),
     tipsPanel: ({ value }) => {
       const textColor = useColorModeValue(
         "text.default.light",
